requirements.txt
requests
urllib3
PyYAML
loguru
